Move permission sets back to app/
These were already in app, but moved in
https://github.com/solidusio/solidus/pull/403.

At the time of that PR, code reloading was still handled by Rails, so
let's see if we can now move them back as Zeitwerk now handles code
reloading.
