+++ core/lib/spree/core.rb
@@ -115,7 +115,6 @@ require 'spree/core/state_machines'
 require 'spree/core/stock_configuration'
 require 'spree/core/null_promotion_configuration'
 require 'spree/core/validators/email'
-require 'spree/permission_sets'
 require 'spree/user_class_handle'
 
 require 'spree/preferences/store'
