+++ core/lib/spree/permission_sets.rb
@@ -1,18 +1,7 @@
 # frozen_string_literal: true
 
-require 'spree/permission_sets/base'
-require 'spree/permission_sets/configuration_display'
-require 'spree/permission_sets/configuration_management'
-require 'spree/permission_sets/dashboard_display'
-require 'spree/permission_sets/default_customer'
-require 'spree/permission_sets/order_display'
-require 'spree/permission_sets/order_management'
-require 'spree/permission_sets/product_display'
-require 'spree/permission_sets/product_management'
-require 'spree/permission_sets/restricted_stock_display'
-require 'spree/permission_sets/restricted_stock_management'
-require 'spree/permission_sets/stock_display'
-require 'spree/permission_sets/stock_management'
-require 'spree/permission_sets/super_user'
-require 'spree/permission_sets/user_display'
-require 'spree/permission_sets/user_management'
+Spree.deprecator.warn(
+  <<~MSG
+    The file "#{__FILE__}" does not need to be `require`d any longer, it is now autoloaded.
+  MSG
+)
