+++ core/lib/spree/permission_sets/stock_management.rb
@@ -1,26 +1,7 @@
 # frozen_string_literal: true
 
-module Spree
-  module PermissionSets
-    # Full permissions for stock management.
-    #
-    # This permission set grants full control over all stock items and read
-    # access to locations.
-    class StockManagement < PermissionSets::Base
-      class << self
-        def privilege
-          :management
-        end
-
-        def category
-          :stock
-        end
-      end
-
-      def activate!
-        can :manage, Spree::StockItem
-        can :read, Spree::StockLocation
-      end
-    end
-  end
-end
+Spree.deprecator.warn(
+  <<~MSG
+    The file "#{__FILE__}" does not need to be `require`d any longer, it is now autoloaded.
+  MSG
+)
